+++ src/main/kotlin/de/bdr/servko/keycloak/gematik/idp/validation/GematikIdpCertificateValidatorSpi.kt
@@ -1,0 +1,22 @@
+package de.bdr.servko.keycloak.gematik.idp.validation
+
+import org.keycloak.provider.Provider
+import org.keycloak.provider.ProviderFactory
+import org.keycloak.provider.Spi
+
+class GematikIdpCertificateValidatorSpi : Spi {
+
+    companion object {
+        const val NAME = "gematik-idp-certificate-validator"
+    }
+
+    override fun isInternal(): Boolean = false
+
+    override fun getName(): String = NAME
+
+    override fun getProviderClass(): Class<out Provider> =
+        GematikIdpCertificateValidatorProvider::class.java
+
+    override fun getProviderFactoryClass(): Class<out ProviderFactory<out Provider>> =
+        GematikIdpCertificateValidatorProviderFactory::class.java
+}