+++ src/test/kotlin/de/bdr/servko/keycloak/gematik/idp/rest/GematikIDPResourceTest.kt
@@ -20,43 +20,29 @@ package de.bdr.servko.keycloak.gematik.idp.rest
 import de.bdr.servko.keycloak.gematik.idp.GematikIDP
 import de.bdr.servko.keycloak.gematik.idp.model.AuthenticationFlowType
 import de.bdr.servko.keycloak.gematik.idp.model.GematikIDPConfig
+import de.bdr.servko.keycloak.gematik.idp.validation.GematikIdpCertificateValidatorProvider
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import org.keycloak.broker.provider.IdentityProvider
+import org.keycloak.broker.provider.UserAuthenticationIdentityProvider
 import org.keycloak.forms.login.LoginFormsProvider
 import org.keycloak.models.KeycloakSession
 import org.keycloak.models.RealmModel
+import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 
 internal class GematikIDPResourceTest {
 
     private val realmMock = mock<RealmModel>()
-    private val callbackMock = mock<IdentityProvider.AuthenticationCallback>()
-    private val sessionMock = mock<KeycloakSession> ()
+    private val callbackMock = mock<UserAuthenticationIdentityProvider.AuthenticationCallback>()
+    private val certificateValidator = mock<GematikIdpCertificateValidatorProvider>()
+    private val sessionMock = mock<KeycloakSession> () {
+        on { getProvider(GematikIdpCertificateValidatorProvider::class.java) } doReturn certificateValidator
+    }
     private val idp = mock<GematikIDP> ()
     private val formsMock = mock<LoginFormsProvider> ()
     private val config = mock<GematikIDPConfig>()
 
-
-    @Test
-    fun fromConfigWithLegacyFlow_ReturnsLegacyEndpoint() {
-        // arrange
-        whenever(config.getAuthenticationFlow()).thenReturn(AuthenticationFlowType.LEGACY)
-
-        // act
-        val result = GematikIDPResource.from(
-            realm = realmMock,
-            callback = callbackMock,
-            session = sessionMock,
-            gematikIDP = idp,
-            config = config,
-            forms = formsMock
-        )
-
-        // assert
-        assertThat(result).isInstanceOf(GematikIDPLegacyResource::class.java)
-    }
 
     @Test
     fun fromConfigWithMultiFlow_ReturnsMultiEndpoint() {
@@ -70,7 +56,7 @@ internal class GematikIDPResourceTest {
             session = sessionMock,
             gematikIDP = idp,
             config = config,
-            forms = formsMock
+            forms = formsMock,
         )
 
         // assert
@@ -89,7 +75,7 @@ internal class GematikIDPResourceTest {
             session = sessionMock,
             gematikIDP = idp,
             config = config,
-            forms = formsMock
+            forms = formsMock,
         )
 
         // assert
@@ -108,7 +94,7 @@ internal class GematikIDPResourceTest {
             session = sessionMock,
             gematikIDP = idp,
             config = config,
-            forms = formsMock
+            forms = formsMock,
         )
 
         // assert